+++ qt project/NievaTrains_P2/model/gerarchia/electric.h
@@ -12,17 +12,17 @@ protected:
 private:
     TtrasmissioneElettrico trasmissioneElettrico;
 protected:
-    float efficenzaElettrico;
+    double efficenzaElettrico;
 public:
-    Electric(const std::string& = "NoName", const std::string& ="NoBuilder", unsigned int =100, Trotaia =Trotaia::maglev, Ttreno =Ttreno::alta_velocita,  TtrasmissioneElettrico =TtrasmissioneElettrico::overhead_lines, float =0.7f);
+    Electric(const std::string& = "NoName", const std::string& ="NoBuilder", unsigned int =100, Trotaia =Trotaia::maglev, Ttreno =Ttreno::alta_velocita,  TtrasmissioneElettrico =TtrasmissioneElettrico::overhead_lines, double =0.7);
 
     std::string getTrasmissioneElettrico() const;
-    float getEfficenzaElettrico() const;
+    double getEfficenzaElettrico() const;
 
     void setTrasmissioneElettrico(std::string);
-    void setEfficenzaElettrico(float);
+    void setEfficenzaElettrico(double);
 
-    float carburanteNecessario(unsigned int) const;  //input km da percorrere
+    double carburanteNecessario(unsigned int) const;  //input km da percorrere
     unsigned int kmPercorribili(unsigned int) const; //input kw di carburante
     std::string type() const;
     void print() const;
